Server error component and troubleshot show under main info instead of switch
