+++ app/assets/styles/base/forms.scss
@@ -182,48 +182,58 @@ bex-create-form {
         }
     }
 
-    > .error-banner {
-        background: $red;
-        color: $whiteSmoke;
-        padding: 10px;
-        border-radius: 4px;
+}
+
+.error-banner {
+    background: $red;
+    color: $whiteSmoke;
+    border-radius: 4px;
+
+    > .content {
         display: flex;
+        position: relative;
 
-        > .content {
-            flex: 1;
-            position: relative;
-            overflow: hidden;
-            padding-right: 10px;
-            display: flex;
-
-            h4, p, table {
-                color: $whiteSmoke;
-                font-size: 1em;
-            }
-
-            h4 {
-                font-weight: 500;
-            }
-
-            td {
-                padding: 0 5px;
-            }
-
+        h4, p, table {
+            color: $whiteSmoke;
+            font-size: 1em;
         }
 
-        > .troubleshoot-btn {
-            font-size: 1.5em;
-            cursor: pointer;
-            z-index: 10;
+        h4 {
+            font-weight: 500;
+        }
 
-            i {
-                border: 1px solid $red-dark;
-                border-radius: 3px;
-                padding: 3px;
-                &:hover {
-                    background-color: $red-dark;
-                }
+        .main-info {
+            flex: 1;
+            padding: 10px;
+        }
+    }
+
+    .troubleshoot-btn {
+        font-size: 1.5em;
+        cursor: pointer;
+        z-index: 10;
+
+        i {
+            border: 1px solid $red-dark;
+            border-radius: 3px;
+            padding: 3px;
+            margin: 10px;
+            &:hover {
+                background-color: $red-dark;
             }
+        }
+    }
+
+    .troubleshoot-info {
+        padding: 10px;
+        border-top: 1px solid $red-dark;
+
+        table {
+            font-size: 1em;
+            color: $whiteSmoke;
+        }
+        td {
+            padding-right: 10px;
         }
     }
 }
