+++ app/components/base/form/server-error/index.ts
@@ -1,0 +1,1 @@
+export * from "./server-error.component";
