+++ app/components/base/form/server-error/server-error.html
@@ -1,0 +1,26 @@
+<div class="error-banner" *ngIf="error">
+    <div class="content">
+        <div class="main-info">
+            <h4>{{error.code}}</h4>
+            <p>{{errorData.message}}</p>
+            <div class="details" *ngIf="error.values">
+                <div *ngFor="let detail of error.values">{{detail.key}}: {{detail.value}}</div>
+            </div>
+        </div>
+        <div class="troubleshoot-btn">
+            <i class="fa fa-bug" (click)="showErrorTroubleshoot = !showErrorTroubleshoot" title="Troubleshoot details"></i>
+        </div>
+    </div>
+    <div class="troubleshoot-info" *ngIf="showErrorTroubleshoot">
+        <table>
+            <tr>
+                <td><b>RequestId</b></td>
+                <td>{{errorData.requestId}}</td>
+            </tr>
+            <tr>
+                <td><b>Time</b> </td>
+                <td>{{errorData.time}}</td>
+            </tr>
+        </table>
+    </div>
+</div>
