+++ app/components/pool/action/add/pool-create-basic-dialog.html
@@ -1,6 +1,6 @@
 <div class="slideout-form content-wrapper">
     <h1>Add a pool to the account</h1>
-    <p>Pools allow you to provision worker nodes to do work</p>
+    <p>Pools allow you to provision compute nodes to do work</p>
     <bex-create-form [formGroup]="createPoolForm" [submit]="submit" [sidebarRef]="sidebarRef">
         <form novalidate [formGroup]="createPoolForm">
             <div class="form-element">
